% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ancestry.strat.test.R
\name{ancestry.strat.test}
\alias{ancestry.strat.test}
\title{A function to test if there is a difference between old and new AHA classifications, stratified by ancestry}
\usage{
ancestry.strat.test(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input that includes the variables SBP_meas, DBP_meas, 
and hypertension variables created using the function "htn_outcome_classify"}
}
\value{
Will return a new list with 5 objects:
   \item{testSummary}{Summary table of P-values for all 4 tests}
   \item{stratSumAfr}{Contingency table of old vs new AHA 3-level in African ancestry}
   \item{stratSumEur}{Contingency table of old vs new AHA 3-level in European ancestry}
   \item{stratSumAfrBi}{Contingency table of old vs new AHA 2-level in African ancestry}
   \item{stratSumEurBi}{Contingency table of old vs new AHA 2-level in European ancestry}
}
\description{
This function allows you to test two- (non-hypertensive vs. hypertensive)
and three-level hypertension (normotensive, pre-hypertensive, hypertensive)
classifications to see if there is a significant difference in classification
based on old and new AHA guidelines.
The tests are stratified by ancestry.
}
\details{
OLD AHA BP CUTOFFS:
   Normotensive: SBP <120 mm Hg AND DBP <80 mm Hg
   Pre-hypertensive: SBP 120-139 mm Hg OR DBP 80-89 mm Hg
   Hypertensive: SBP ≥140 mm Hg OR DBP ≥90 mm Hg

NEW AHA BP CUTOFFS:
   Normotensive: SBP <120 mm Hg AND DBP <80 mm Hg
   Pre-hypertensive/Elevated: SBP 120-129 mm Hg AND DBP <80 mm Hg
   Hypertensive: SBP ≥130 mm Hg OR DBP ≥80 mm Hg
}
\examples{
STUDYPI_ancestry_strat_test <- ancestry.strat.test(classed)
}
\keyword{Fisher's}
\keyword{ancestry}
\keyword{classification,}
\keyword{exact}
\keyword{hypertension}
\keyword{stratification,}
