% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.PRS.cutoff.R
\name{find.PRS.cutoff}
\alias{find.PRS.cutoff}
\title{A Function to run identify best PRS cutoff}
\usage{
find.PRS.cutoff(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input}
}
\value{
Returns plot of R-squared values for each PRS cutoff
}
\description{
This function creates an easy-to-interpret summary to select the best PRS cutoff 
for use in the main step of the regression analysis
}
\examples{
polygenicPTSD_contin(classed)
}
\keyword{polygenic}
\keyword{risk}
\keyword{score}
