% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htn_outcome_classify.R
\name{htn_outcome_classify}
\alias{htn_outcome_classify}
\title{A Function to Classify continuous SBP and DBP outcomes as categorical hypertension}
\usage{
htn_outcome_classify(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input that includes the variables SBP_meas, DBP_meas, and antihtn_use -0 or 1-}
}
\description{
This function allows you to create new categorical hypertension variables in your dataset based on old and new AHA guidelines
}
\examples{
htn_outcome_classify(merged)
}
\keyword{DBP,}
\keyword{SBP,}
\keyword{classification}
\keyword{hypertension,}
