% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenicPTSD_contin.R
\name{polygenicPTSD_contin}
\alias{polygenicPTSD_contin}
\title{A Function to run analysis on continuous BP measures}
\usage{
polygenicPTSD_contin(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input}
}
\value{
Will return regression objects for all required models, 
and save them in your working directory. 

Specifically, runs main effect and interaction models for all 
4 PTSD variables and PRS variables, using a base model of sex, age, first three 
ancestry principle components. 

Regression is stratified by ancestry, and sensitivity analyses
performed for gender, education, and trauma count.
}
\description{
This function allows you to quickly run and save the required regression
analyses for the pipeline.
}
\examples{
polygenicPTSD_contin(classed)
}
\keyword{continuous}
\keyword{linear}
\keyword{outcome}
\keyword{regression,}
