% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenicPTSD_logit.R
\name{polygenicPTSD_logit}
\alias{polygenicPTSD_logit}
\title{A function to run logistic regression on binary hypertension outcome}
\usage{
polygenicPTSD_logit(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input}
}
\value{
Will return glm objects for all required models, 
and save them in your working directory. 

Specifically, runs main effect and interaction models for all 
4 PTSD variables and PRS variables, using a base model of sex, age, first three 
ancestry principle components. Outcome is binary hypertension outcome
(non-hypertensive vs hypertensive).

Regression is stratified by ancestry, and sensitivity analyses
are performed for gender, education, and trauma count.
}
\description{
This function allows you to quickly run and save the required logistic
models for the analysis pipeline.
}
\examples{
polygenicPTSD_logit(classed)
}
\keyword{categorical}
\keyword{logistic}
\keyword{outcome}
\keyword{regression,}
