% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygenicPTSD_multinom.R
\name{polygenicPTSD_multinom}
\alias{polygenicPTSD_multinom}
\title{A function to run multinomial regression on categorical hypertension outcome}
\usage{
polygenicPTSD_multinom(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input}
}
\value{
Will return multinom objects for all required models, 
and save them in your working directory. 

Specifically, runs main effect and interaction models for all 
4 PTSD variables and PRS variables, using a base model of sex, age, first three 
ancestry principle components. Outcome is 3-level categorical hypertension outcome
(normotensive, pre-hypertensive, and hypertensive).

Regression is stratified by ancestry, and sensitivity analyses
are performed for gender, education, and trauma count.
}
\description{
This function allows you to quickly run and save the required multinomial
models for the analysis pipeline.
}
\examples{
polygenicPTSD_multinom(classed)
}
\keyword{categorical}
\keyword{multinomial}
\keyword{outcome}
\keyword{regression,}
