% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prsVars.R
\name{prsVars}
\alias{prsVars}
\title{A Function to grab PRS and PTSD variables from the dataset}
\usage{
prsVars(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input that includes the variables for SBP PRS (starting with "sbp_prs"), DBP PRS (starting with "dbp_prs") and Hypertension PRS (starting with "bp_prs")}
}
\description{
This function allows you to create new categorical hypertension variables in your dataset based on old and new AHA guidelines
}
\examples{
prsVars(classed)
}
\keyword{DBP,}
\keyword{PRS}
\keyword{SBP,}
