% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable.summary.stratified.R
\name{variable.summary.stratified}
\alias{variable.summary.stratified}
\title{A function to save summary statistics to working directory}
\usage{
variable.summary.stratified(data_entry)
}
\arguments{
\item{data_entry}{Dataframe input that includes all required variables}
}
\value{
Will return two summaries (one for each ancestry) and save those summaries to your working directory
   \item{STUDYPI_summary_vars_index_afr}{Summary statistics for varibles in the African ancestry subset}
   \item{STUDYPI_summary_vars_index_eur}{Summary statistics for varibles in the European ancestry subset}
}
\description{
This function allows you to easily save summary/descriptive statistics of variables
stratified by ancestry
}
\examples{
STUDYPI_ancestry_strat_test <- ancestry.strat.test(classed)
}
\keyword{ancestry}
\keyword{stratification}
\keyword{summary,}
